python_server/requirements.txt
llama-index-core==0.12.28
llama-index-llms-openai==0.3.30
llama-index-embeddings-openai==0.3.1
openai>=1.1.1,<2.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
jinja2==3.1.3
# User authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
python-multipart>=0.0.5
# Database (Added for PostgreSQL Integration)
sqlalchemy>=2.0
psycopg2-binary
alembic 